Port mathex functionality to kernel space
I use self-defined floating point format to implement floating point in
kernel mode. The 32 bits of "int" is divided into 28 bits of mantissa
and 4 bits of exponent.

The feature for user to send their self-defined function for calculation
is not implemented. Other functionalities should be well except for
the precision problem.

The precision for this version is now less than 9 digits since I use
only 28 bits to save the number.